+++ src/components/Sort/index.js
@@ -1,37 +1,18 @@
-import {
-  MenuItem,
-  FormControl,
-  Select,
-  Typography,
-  InputBase,
-  Button,
-  SvgIcon as MuiSvgIcon,
-} from "@material-ui/core";
+import { FormControl, Typography, Button } from "@material-ui/core";
 import Image from "next/image";
 import PropTypes from "prop-types";
 import React from "react";
 
 import useStyles from "./useStyles";
 
-import { ReactComponent as KeyboardArrowDown } from "@/twoopstracker/assets/icons/select-dropdown.svg";
 import SortDown from "@/twoopstracker/assets/icons/sort-down.svg";
 import SortUp from "@/twoopstracker/assets/icons/sort-up.svg";
-
-function SvgIcon(props) {
-  const classes = useStyles(props);
-  return (
-    <MuiSvgIcon
-      component={KeyboardArrowDown}
-      className={classes.svgSelectIcon}
-    />
-  );
-}
+import TweetSelect from "@/twoopstracker/components/TweetSelect";
 
 function Sort({
   handleSelection,
   toggleSortOrder,
   isDesc,
-  children,
   label,
   name,
   menuItems,
@@ -52,32 +33,13 @@ function Sort({
       <Typography variant="body1" className={classes.label}>
         {label}
       </Typography>
-      <Select
-        labelId={`${label}-id`}
+      <TweetSelect
+        name={name}
+        label={label}
         value={value}
-        defaultValue={value}
-        onChange={handleChange}
-        className={classes.select}
-        IconComponent={SvgIcon}
-        placeholder={value}
-        input={
-          <InputBase
-            id={`${label}-id`}
-            inputProps={{ "aria-label": `${label}-id` }}
-            classes={{
-              root: classes.inputBase,
-              input: classes.inputBaseInput,
-            }}
-          />
-        }
-      >
-        {menuItems &&
-          menuItems.map((item) => (
-            <MenuItem key={item.value} value={item.value}>
-              {item.name}
-            </MenuItem>
-          ))}
-      </Select>
+        handleChange={handleChange}
+        menuItems={menuItems}
+      />
       <Button
         color="default"
         variant="text"
@@ -102,7 +64,6 @@ Sort.propTypes = {
   toggleSortOrder: PropTypes.func,
   isDesc: PropTypes.bool,
   handleSelection: PropTypes.func,
-  children: PropTypes.node,
   name: PropTypes.string,
   label: PropTypes.string,
   menuItems: PropTypes.arrayOf(
@@ -117,7 +78,6 @@ Sort.propTypes = {
 Sort.defaultProps = {
   toggleSortOrder: undefined,
   handleSelection: undefined,
-  children: undefined,
   label: undefined,
   name: undefined,
   menuItems: undefined,
