+++ src/components/TweetSelect/index.stories.js
@@ -1,0 +1,15 @@
+import React from "react";
+
+import TweetSelect from ".";
+
+export default {
+  title: "Components/TweetSelect",
+};
+
+function Template() {
+  return <TweetSelect />;
+}
+
+export const Default = Template.bind({});
+
+Default.args = {};
